fix: dead loop
