+++ docs/examples/handle.tsx
@@ -1,81 +1,7 @@
-import 'rc-tooltip/assets/bootstrap.css';
 import React from 'react';
 import Slider from 'rc-slider';
-import type { SliderProps } from 'rc-slider';
-import raf from 'rc-util/lib/raf';
-import Tooltip from 'rc-tooltip';
 import '../../assets/index.less';
-
-const HandleTooltip = (props: {
-  value: number;
-  children: React.ReactElement;
-  visible: boolean;
-  tipFormatter?: (value: number) => React.ReactNode;
-}) => {
-  const { value, children, visible, tipFormatter = (val) => `${val} %` } = props;
-
-  const tooltipRef = React.useRef<any>();
-  const rafRef = React.useRef<number | null>(null);
-
-  function cancelKeepAlign() {
-    raf.cancel(rafRef.current!);
-  }
-
-  function keepAlign() {
-    rafRef.current = raf(() => {
-      tooltipRef.current?.forcePopupAlign();
-    });
-  }
-
-  React.useEffect(() => {
-    if (visible) {
-      keepAlign();
-    } else {
-      cancelKeepAlign();
-    }
-
-    return cancelKeepAlign;
-  }, [value, visible]);
-
-  return (
-    <Tooltip
-      placement="top"
-      overlay={tipFormatter(value)}
-      overlayInnerStyle={{ minHeight: 'auto' }}
-      ref={tooltipRef}
-      visible={visible}
-    >
-      {children}
-    </Tooltip>
-  );
-};
-
-const handleRender: SliderProps['handleRender'] = (node, props) => {
-  return (
-    <HandleTooltip value={props.value} visible={props.dragging}>
-      {node}
-    </HandleTooltip>
-  );
-};
-
-const TooltipSlider = ({
-  tipFormatter,
-  ...props
-}: SliderProps & { tipFormatter?: (value: number) => React.ReactNode }) => {
-  const tipHandleRender: SliderProps['handleRender'] = (node, handleProps) => {
-    return (
-      <HandleTooltip
-        value={handleProps.value}
-        visible={handleProps.dragging}
-        tipFormatter={tipFormatter}
-      >
-        {node}
-      </HandleTooltip>
-    );
-  };
-
-  return <Slider {...props} handleRender={tipHandleRender} />;
-};
+import TooltipSlider, { handleRender } from './components/TooltipSlider';
 
 const wrapperStyle = { width: 400, margin: 50 };
